requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.1.2
Django==3.2.12
django-crontab==0.7.1
djangorestframework==3.13.1
idna==3.3
libretranslatepy==2.1.1
lxml==4.8.0
pytz==2022.1
requests==2.27.1
sqlparse==0.4.2
translate==3.6.1
urllib3==1.26.9
